+++ package.json
@@ -2,6 +2,7 @@
   "name": "@geekjuice/verse",
   "version": "1.0.0",
   "main": "index.js",
+  "description": "CLI for Genius lyrics",
   "repository": "geekjuice/verse",
   "author": {
     "name": "Nicholas Hwang",
@@ -12,12 +13,12 @@
   "engines": {
     "node": ">=8.0.0"
   },
-  "bin": "lib/cli.js",
+  "bin": "./dist/cli.js",
   "files": [
-    "lib"
+    "dist/**/*.js"
   ],
   "dependencies": {
-    "chalk": "2.0.0",
+    "chalk": "2.0.1",
     "debug": "2.6.8",
     "dotenv": "4.0.0",
     "immutable": "3.8.1",
@@ -28,14 +29,24 @@
     "ora": "1.3.0"
   },
   "devDependencies": {
+    "@types/chalk": "0.4.31",
+    "@types/debug": "0.0.29",
+    "@types/inquirer": "0.0.35",
+    "@types/keytar": "4.0.0",
+    "@types/minimist": "1.2.0",
+    "@types/node": "^8.0.7",
+    "@types/request": "0.0.45",
     "babel-eslint": "7.2.3",
     "eslint": "4.1.1",
-    "eslint-config-prettier": "2.2.0",
-    "husky": "0.14.1",
+    "eslint-config-prettier": "2.3.0",
+    "husky": "0.14.2",
     "lint-staged": "4.0.0",
-    "prettier": "1.5.2"
+    "prettier": "1.5.2",
+    "typescript": "2.4.1",
+    "typescript-eslint-parser": "3.0.0"
   },
   "scripts": {
+    "tsc": "tsc",
     "eslint": "eslint",
     "eslint:run": "npm run eslint -- lib",
     "prettier": "prettier --write --single-quote",
@@ -43,14 +54,18 @@
     "precommit": "lint-staged"
   },
   "lint-staged": {
-    "*.js": [
+    "*.ts": [
+      "npm run tsc",
       "npm run prettier",
       "npm run eslint",
       "git add"
     ]
   },
   "eslintConfig": {
-    "parser": "babel-eslint",
+    "parser": "typescript-eslint-parser",
+    "parserOptions": {
+      "sourceType": "module"
+    },
     "env": {
       "browser": true,
       "es6": true,
