+++ tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "compilerOptions": {
+    "target": "esnext",
+    "module": "commonjs",
+    "outDir": "dist",
+    "sourceMap":true,
+    "allowSyntheticDefaultImports": true
+  },
+  "files": [
+    "node_modules/immutable/dist/immutable.d.ts"
+  ],
+  "include": [
+    "lib/**/*.ts",
+    "types/**/*.d.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
