+++ types/nightmare/index.d.ts
@@ -1,0 +1,6 @@
+// Type definitions for nightmare 2.10.0
+// Project: https://github.com/segmentio/nightmare
+// Definitions by: Nicholas Hwang <http://nicholash.wang>
+
+declare function nightmare(): any
+export = nightmare;
